+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# Changelog v0.5.5
+Not Released Yet
+
+## Content
+- New notifications system to alert when there are new Reviews in queue (can be disabled)
+
+## Popup
+- Improved "romaji to kana" system. Now, writing, for example, "kko" becomes "っこ", as it should
+
+## Bug Fix
+- Fixed Kanji Search Settings Navbar taking time to show up, or not loading at all
+- Details Popup no longer closes for no reason when clicking to see the details popup of a subject from within that details popup
+
 # [Changelog v0.5.4](https://github.com/digas99/wanikani-kanji-highlighter/releases/tag/v0.5.4)
 Released on 03/09/2021